Move the bulk of the dev API impl into its own package
This will allow it to be called by other tooling, in addition to the developer API
